requirements.txt
Pillow==9.0.1
Pillow==10.3.0
Requests==2.32.3
streamlit==1.35.0
